refactor(*.py): Extract the module for YahooFurigana API. fix #477 @1.5h
