+++ tests/test_yahooapis.py
@@ -1,0 +1,29 @@
+# -*- coding: UTF-8 -*-
+
+# Copyright 2018 [name of copyright owner]
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+import unittest
+
+import yahooapis
+
+
+class TestInfra(unittest.TestCase):
+    def test_replace_hiragana_to_katakana(self):
+        assert yahooapis.replace_hiragana_to_katakana(u'ふりがな') == u'フリガナ'
+        assert yahooapis.replace_hiragana_to_katakana(u'りんご') == u'リンゴ'
